Fix DRBG generation algorithm & guessing

+++ src/curves.rs
@@ -5,13 +5,13 @@ use math::{mod_inverse, prime_mod_inverse};
 
 #[derive(Clone, Debug, PartialEq)]
 pub struct Curve {
-    pub name: String,   // Name of the curve
-    pub bitsize: u16,   // Level of security offered by the curve 
-    pub p: BigInt,      // Order of the finite field that the curve belongs to (mod p)
-    pub n: BigInt,      // Order of the elliptic curve group
-    pub a: BigInt,      // a parameter in the curve equation
-    pub b: BigInt,      // b parameter in the curve equation
-    pub g: CurvePoint   // Generator (base) point
+    pub name: String,       // Name of the curve
+    pub bitsize: usize,     // Level of security offered by the curve 
+    pub p: BigInt,          // Order of the finite field that the curve belongs to (mod p)
+    pub n: BigInt,          // Order of the elliptic curve group
+    pub a: BigInt,          // a parameter in the curve equation
+    pub b: BigInt,          // b parameter in the curve equation
+    pub g: CurvePoint       // Generator (base) point
 }
 
 impl Curve {
